requirements.txt
asgiref==3.5.2
beautifulsoup4==4.12.2
bs4==0.0.1
decorator==5.1.1
Django==4.1.2
django-allauth==0.52.0
django-crispy-forms==1.14.0
django-filter==22.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
importlib-metadata==4.6.4
ipython==8.7.0
oauthlib==3.2.0
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3